o1/requirements.txt
pdfplumber 
spacy 
sentence-transformers 
faiss-cpu 
psutil 
watchdog 
pytest
numpy
Pillow
pytesseract
pymupdf
PyPDF2

# python -m spacy download en_core_web_sm